Fix build errors
Run rustfmt

Fix some error from rebasing

Fix some wrong imports

Remove unused embed_migrations macro from diesel_infer_schema

Add missing dev-dependency to migrations_internals

Add missing -- to .travis.yaml

Add missing dev-dependencies for diesel_migrations

More stuff missing from rebase

Run rustfmt

Remove unused dependency

Remove accidentally set default feature

Remove unused macro_use annotation

Sqlite tests need macro import for diesel_migrations
