+++ diesel_cli/tests/tests.rs
@@ -4,6 +4,7 @@ extern crate chrono;
 extern crate diesel;
 #[macro_use]
 extern crate difference;
+extern crate migrations_internals;
 extern crate regex;
 extern crate tempdir;
