+++ diesel_derives/tests/tests.rs
@@ -3,7 +3,7 @@ extern crate cfg_if;
 #[macro_use]
 extern crate diesel;
 #[macro_use]
-extern crate diesel_codegen;
+extern crate diesel_derives;
 
 mod queryable;
 mod queryable_by_name;
