+++ diesel_migrations/migrations_macros/Cargo.toml
@@ -13,6 +13,9 @@ migrations_internals = { version = "0.16.0", path = "../migrations_internals" }
 syn = { version = "0.11.4", features = ["aster"] }
 quote = "0.3.12"
 
+[dev-dependencies]
+tempdir = "0.3.4"
+
 [lib]
 proc-macro = true
