+++ diesel_migrations/migrations_macros/src/lib.rs
@@ -13,11 +13,11 @@
                   unicode_not_nfc, enum_glob_use, if_not_else, items_after_statements,
                   used_underscore_binding))]
 #![cfg_attr(all(test, feature = "clippy"), allow(option_unwrap_used, result_unwrap_used))]
+extern crate migrations_internals;
 extern crate proc_macro;
 #[macro_use]
 extern crate quote;
 extern crate syn;
-extern crate migrations_internals;
 
 mod migrations;
 mod embed_migrations;
