+++ diesel_migrations/src/lib.rs
@@ -9,10 +9,10 @@
            warn(wrong_pub_self_convention, mut_mut, non_ascii_literal, similar_names,
                   unicode_not_nfc, if_not_else, items_after_statements, used_underscore_binding))]
 
+extern crate migrations_internals;
 #[allow(unused_imports)]
 #[macro_use]
 extern crate migrations_macros;
-extern crate migrations_internals;
 #[doc(hidden)]
 pub use migrations_macros::*;
 #[doc(inline)]
