--- diesel_proc_macro_internal/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "diesel_proc_macro_internal"
-version = "0.16.0"
-authors = ["Sean Griffin <sean@seantheprogrammer.com>"]
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-syn = { version = "0.11.4", features = ["aster"] }
-quote = "0.3.12"
-dotenv = { version = ">=0.8, <0.11", optional = true, default-features = false }
-diesel = { version = "0.16.0", default-features = false }
-diesel_infer_schema = { version = "0.16.0", default-features = false }
-clippy = { optional = true, version = "=0.0.162" }
-
-[dev-dependencies]
-tempdir = "^0.3.4"
-
-[lib]
-proc-macro = true
-
-[features]
-default = ["dotenv", "sqlite"]
-lint = ["clippy"]
-postgres = ["diesel/postgres", "diesel_infer_schema/postgres"]
-sqlite = ["diesel/sqlite", "diesel_infer_schema/sqlite"]
-mysql = ["diesel/mysql", "diesel_infer_schema/mysql"]
