+++ examples/postgres/all_about_updates/Cargo.toml
@@ -5,4 +5,3 @@ authors = ["Sean Griffin <sean@seantheprogrammer.com>"]
 
 [dependencies]
 diesel = { version = "0.16.0", features = ["postgres"] }
-diesel_infer_schema = { path = "../../../diesel_infer_schema", version = "0.16.0" }
